+++ config/development.yaml
@@ -21,11 +21,11 @@ services:
   citizen:
     uploadDocsEmail: UPLOAD_DOCUMENTS_EMAIL
   ccase:
-    url: 'https://ccd-data-store-api-prl-ccd-definitions-pr-1928.preview.platform.hmcts.net'
+    url: 'https://ccd-data-store-api-prl-ccd-definitions-pr-2009.preview.platform.hmcts.net'
   cos:
-    url: 'https://prl-cos-pr-2196.preview.platform.hmcts.net'
+    url: 'https://prl-cos-pr-2205.preview.platform.hmcts.net'
   documentManagement:
-    url: 'https://ccd-case-document-am-api-prl-ccd-definitions-pr-1928.preview.platform.hmcts.net'
+    url: 'https://ccd-case-document-am-api-prl-ccd-definitions-pr-2009.preview.platform.hmcts.net'
   health:
   timeout: 5000
   deadline: 15000
